third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor_advance_index2-expected.txt

This is a testharness.js-based test.
PASS IDBCursor.advance() - attempt to pass a count parameter that is not a number 
Harness: the test ran to completion.
